+++ stor/tests/test_posix.py
@@ -1,4 +1,3 @@
-import mock
 import os
 import tempfile
 import unittest
@@ -7,9 +6,6 @@ import stor
 from stor import NamedTemporaryDirectory
 from stor import Path
 from stor import posix
-from stor_s3 import s3
-from stor import settings
-from stor_swift import swift
 from stor import utils
 from stor import windows
 
@@ -28,20 +24,6 @@ class TestDiv(unittest.TestCase):
         with self.assertRaisesRegexp(TypeError, 'unsupported operand'):
             posix.PosixPath('my/path') / windows.WindowsPath(r'other\path')
 
-    def test_w_swift_component(self):
-        p = posix.PosixPath('my/path') / swift.SwiftPath('swift://t/c/name').name
-        self.assertEquals(p, posix.PosixPath('my/path/name'))
-        self.assertEquals(stor.join('my/path',
-                                    swift.SwiftPath('swift://t/c/name').name),
-                          p)
-
-    def test_w_s3_component(self):
-        p = posix.PosixPath('my/path') / s3.S3Path('s3://b/name').name
-        self.assertEquals(p, posix.PosixPath('my/path/name'))
-        self.assertEquals(stor.join('my/path',
-                                    s3.S3Path('s3://b/name').name),
-                          p)
-
 
 class TestAdd(unittest.TestCase):
     def test_success(self):
@@ -51,14 +33,6 @@ class TestAdd(unittest.TestCase):
     def test_w_windows_path(self):
         with self.assertRaisesRegexp(TypeError, 'unsupported operand'):
             posix.PosixPath('my/path') + windows.WindowsPath(r'other\path')
-
-    def test_w_swift_component(self):
-        p = posix.PosixPath('my/path') + swift.SwiftPath('swift://t/c/name').name
-        self.assertEquals(p, posix.PosixPath('my/pathname'))
-
-    def test_w_s3_component(self):
-        p = posix.PosixPath('my/path') + s3.S3Path('s3://b/name').name
-        self.assertEquals(p, posix.PosixPath('my/pathname'))
 
     def test_invalid_radd(self):
         with self.assertRaisesRegexp(TypeError, 'unsupported operand'):
@@ -123,26 +97,6 @@ class TestCopy(unittest.TestCase):
             with self.assertRaisesRegexp(ValueError, 'copy to tenant'):
                 utils.copy(source / '1', dest)
 
-    @mock.patch.object(swift.SwiftPath, 'upload', autospec=True)
-    def test_swift_destination(self, mock_upload):
-        dest = Path('swift://tenant/container/file.txt')
-        with tempfile.NamedTemporaryFile() as tmp_f:
-            Path(tmp_f.name).copy(dest)
-            upload_args = mock_upload.call_args_list[0][0]
-            self.assertEquals(upload_args[0], dest.parent)
-            self.assertEquals(upload_args[1][0].source, tmp_f.name)
-            self.assertEquals(upload_args[1][0].object_name, 'file.txt')
-
-    @mock.patch.object(s3.S3Path, 'upload', autospec=True)
-    def test_s3_destination(self, mock_upload):
-        dest = Path('s3://bucket/key/file.txt')
-        with tempfile.NamedTemporaryFile() as tmp_f:
-            Path(tmp_f.name).copy(dest)
-            upload_args = mock_upload.call_args_list[0][0]
-            self.assertEquals(upload_args[0], dest.parent)
-            self.assertEquals(upload_args[1][0].source, tmp_f.name)
-            self.assertEquals(upload_args[1][0].object_name, 'key/file.txt')
-
 
 class TestCopytree(unittest.TestCase):
     def test_posix_destination(self):
@@ -182,26 +136,6 @@ class TestCopytree(unittest.TestCase):
 
             with self.assertRaises(OSError):
                 utils.copytree(invalid_source, dest)
-
-    @mock.patch.object(swift.SwiftPath, 'upload', autospec=True)
-    def test_swift_destination(self, mock_upload):
-        source = '.'
-        dest = Path('swift://tenant/container')
-        options = {
-            'swift:upload': {
-                'object_threads': 30,
-                'segment_threads': 40
-            }
-        }
-
-        with settings.use(options):
-            utils.copytree(source, dest)
-        mock_upload.assert_called_once_with(
-            dest,
-            ['.'],
-            condition=None,
-            use_manifest=False,
-            headers=None)
 
 
 class TestOpen(unittest.TestCase):